Migrate player.js to typescript
