+++ src/app/music/MusicRoot.js
@@ -1,5 +1,6 @@
 import { ipcRenderer } from "electron";
-import player from "components/music/helpers/player.js";
+import player from "components/music/helpers/player.ts";
+import { playNotePreview } from "components/music/helpers/notePreview";
 
 const log = (log) => {
   console.log(log);
@@ -80,13 +81,11 @@ ipcRenderer.on("music-data", (event, d) => {
       });
       break;
     case "preview":
-      player.preview(
-        d.note,
-        d.type,
-        d.instrument,
-        d.square2,
-        d.waveForms || []
-      );
+      let waves = d.waveForms || [];
+      if (waves.length === 0) {
+        waves = player.getCurrentSong().waves;
+      }
+      playNotePreview(d.note, d.type, d.instrument, d.square2, waves);
       ipcRenderer.send("music-data-receive", {
         action: "log",
         message: "preview",
